(CDAP-8980) Separate out the Spark class rewrite logic from the SparkRunnerClassLoader
